Move modesl to iolconf models folder and change namespace
