+++ src/PxcIolinkConfiguration/Models/IolConf/Parameters.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Xml.Serialization;
 
-namespace EasyEPlanner.PxcIolinkConfiguration.Models
+namespace EasyEPlanner.PxcIolinkConfiguration.Models.IolConf
 {
     [XmlRoot(ElementName = "Parameters")]
     public class Parameters : ICloneable
